+++ NEWS.md
@@ -1,5 +1,8 @@
 # sjlabelled 1.0.4
 
+## Changes to functions
+* `get_term_labels()` now includes variable names for factors with numeric factor levels only (and not only return the numeric level as term label).
+
 ## Bug fixes
 
 * Fixed bug for `as_label()`, when `x` was a character vector and argument `drop.levels` was `TRUE`.